Added first unit test.
To make the test work from within a folder, I had to put __init__.py
files into the root and the test folder.

Nose is a test manager that seems to be preferred, but I thought
I'd start with unittest, since it's built in.

You have to run the tests from the command line, one directory
above the project (chelmbigstock) folder.

python -m chelmbigstock.tests.stocktests