I added a test for data_download.py. In order to test it, I had to split the URL construction out of the main() code and into a function, so that I could then test that function.
The test checks the URL, *but*, the URL contains today's date,
so it's not entirely possible to check the function against a
static string. I hacked the test to check the 1st and last parts
of the string and to NOT CHECK the middle, today's date of the URL.

With this test working, I plan to continue to refactor data_download
to break out more functions and write more tests against them.